requirements.txt
streamlit 
snowflake-connector-python 
faiss-cpu 
torch 
transformers 
sentence-transformers 
python-dotenv